+++ src/index.ts
@@ -2,8 +2,12 @@ import { StonApiClient, type StonApiClientOptions } from "./client/apiClient";
 
 export { StonApiClient, type StonApiClientOptions };
 
-export { AssetKind } from "./client/types/asset";
+export { AssetKind, AssetTag } from "./client/types/asset";
+export { OperationType } from "./client/types/operation";
 export type AssetInfo = Awaited<ReturnType<StonApiClient["getAssets"]>>[number];
+export type AssetInfoV2 = Awaited<
+  ReturnType<StonApiClient["queryAssets"]>
+>[number];
 export type FarmInfo = Awaited<ReturnType<StonApiClient["getFarms"]>>[number];
 export type FarmNftInfo = FarmInfo["nftInfos"][number];
 export type PoolInfo = Awaited<ReturnType<StonApiClient["getPools"]>>[number];
@@ -11,4 +15,7 @@ export type SwapSimulation = Awaited<ReturnType<StonApiClient["simulateSwap"]>>;
 export type SwapStatus = Awaited<ReturnType<StonApiClient["getSwapStatus"]>>;
 export type SwapPair = Awaited<
   ReturnType<StonApiClient["getSwapPairs"]>
+>[number];
+export type OperationInfo = Awaited<
+  ReturnType<StonApiClient["getOperationsStats"]>
 >[number];