+++ src/utils/normalizeDate.test.ts
@@ -1,0 +1,13 @@
+import { describe, it, expect } from "vitest";
+
+import { normalizeDate } from "./normalizeDate";
+
+describe("normalizeDate", () => {
+  it("should return timestamp in 'YYYY-MM-DDTHH:MM:SS' format from given Date", () => {
+    const timestampString = normalizeDate(
+      new Date("01/19/2024 12:00:00 UTC+0"),
+    );
+
+    expect(timestampString).toBe("2024-01-19T12:00:00");
+  });
+});