+++ src/utils/normalizeDate.ts
@@ -1,0 +1,3 @@
+export const normalizeDate = (date: Date) => {
+  return date.toISOString().split(".")[0];
+};